add getting all ads endpoint with filtering, paging and caching

+++ src/Services/AdService/AdService.Infrastructure.Postgres/Data/Seeding/Fakers/CarSnapshotFaker.cs
@@ -30,7 +30,7 @@ public class CarSnapshotFaker
                 var year = f.Random.Int(generation.YearFrom, generation.YearTo); // >1900
                 var vin = f.Random.String2(CarSnapshot.REQUIRED_VIN_LENGTH, "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789");
                 var mileage = f.Random.Int(0, 500000);
-                var consumption = f.Random.Decimal(5, 50);
+                var consumption = f.Random.Float(5, 50);
                 var color = f.Random.ArrayElement(_colors);
                 var driveType = f.Random.ArrayElement(driveTypes) with { };
                 var transmissionType = f.Random.ArrayElement(transmissionTypes) with { };
